+++ src/screens/WalletScreen/Modals/AvatarModal/AvatarModal.types.ts
@@ -1,0 +1,4 @@
+
+export interface AvatarModalProps {
+	onDismiss: () => void;
+}
